chore: move test to test/components folder
